Move package util.gui -> gui.view.util
