+++ src/main/java/oripa/gui/view/foldability/FoldabilityCheckFrameFactory.java
@@ -6,7 +6,7 @@ import javax.swing.JComponent;
 import javax.swing.JFrame;
 
 import oripa.domain.fold.halfedge.OrigamiModel;
-import oripa.util.gui.ChildFrameManager;
+import oripa.gui.view.util.ChildFrameManager;
 import oripa.value.OriLine;
 
 public class FoldabilityCheckFrameFactory {
