+++ src/main/java/oripa/persistence/doc/exporter/PictureExporter.java
@@ -36,7 +36,7 @@ import oripa.drawer.java2d.Java2DGraphicDrawer;
 import oripa.geom.RectangleDomain;
 import oripa.gui.presenter.creasepattern.CreasePatternGraphicDrawer;
 import oripa.gui.presenter.creasepattern.ObjectGraphicDrawer;
-import oripa.util.gui.AffineCamera;
+import oripa.gui.view.util.AffineCamera;
 
 /**
  * @author Koji
