+++ src/test/java/oripa/persistence/foldformat/CreasePatternElementConverterTest.java
@@ -27,7 +27,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import oripa.persistence.foldformat.CreasePatternElementConverter;
 import oripa.value.OriLine;
 
 /**
